refactor: modulariza a lógica do labirinto em arquivo próprio
